Consolidate sync-orders into sync-orders.ts
Jira ticket: CAMS-421
